Don't deserialize errors
It's tricky to deserialize HTTP errors since they're not guaranteed
to be in a consistent format.

e.g. failure to publish to NSQ returns "Internal Server Error" instead
of a JSON response.

Alternatively, an error message could be served from the load balancer,
NGINX, etc, in which case it most certainly won't be in the same format
as our API.

Additionally, we don't really need to deserialize the errors, since we
don't take any special action.
Simply reading the response and surfacing the raw body along with the
status and status code should be good enough for our purpose.